// src/app/models/image.ts
export class ImageModel {
    UID: number;
    Image1: string;
    Name: string;
    MediaCategory_UID: number;
    GeoLocationLat: string;
    GeoLocationLng: string;
    Description: string;
    Property_UID: number;
    RoomType_UID: number;
    Package_UID: number;
    Attraction_UID: number;
    Extras_UID: number;
    Client_UID: number;
    CreatedDate: any;
    ModifiedDate: any;
    CalenderEvents_UID: number;
    ThirdPartyIntermediary_UID: number;
    IsDeleted: boolean;
    FileExtension: string;
    Revision: string;
    IsFavorite: boolean;
    ImageEncoded: string;
    TrackingId: string;
    ImageData: string;
    height: number;
    width: number;
    Id: number;
    EntityType: any;
    EntityKey: any;
    NewAdded: boolean;

    constructor() {}
}